Add transforms to/from core code block
This adds transforms so that SyntaxHighlighter code blocks can be
directly converted from and to the Gutenberg `core/code` blocks.

When pasting a markdown post into Gutenberg, code blocks in the post are
automatically converted to `core/code` blocks, so I have to manually
change them into SyntaxHighlighter blocks instead. Without this patch,
that means creating a new block, copy/pasting the content of the old
block to the new one, and then deleting the old block. For a long post
this takes quite a lot of time. With this change it will be a two-click
operation per block.